Update estistyles.css
Header button updated. Has a pressed state, a difference in hover vs clicked. NOTICE: to get pressed effect, set the class of the button to pressed for each respective button. As in, the ABOUT page will have its button class set to PRESSED only when the user is on the about page. This is changed in the html manually. Viel spaß!